+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/custom.py
@@ -2409,3 +2409,13 @@ def ssh_webapp(cmd, resource_group_name, name, slot=None):  # pylint: disable=to
         raise CLIError('webapp ssh is only supported on linux and mac')
     else:
         create_tunnel_and_session(cmd, resource_group_name, name, port=None, slot=slot)
+
+
+def create_devops_build(cmd, functionapp_name=None, organization_name=None, project_name=None,
+                        overwrite_yaml=None, use_local_settings=None, local_git=None):
+    from .azure_devops_build_iteractive import AzureDevopsBuildInteractive
+    azure_devops_build_interactive = AzureDevopsBuildInteractive(cmd, logger, functionapp_name,
+                                                                 organization_name, project_name,
+                                                                 overwrite_yaml, use_local_settings,
+                                                                 local_git)
+    return azure_devops_build_interactive.interactive_azure_devops_build()