+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -3,6 +3,7 @@
 Release History
 ===============
 * functionapp: add ability to create and configure functions using ACR containers
+* functionapp: az functionapp devops-build, new command created
 
 0.2.13
 ++++++